changes: refactor framework

+++ D3D12DrawMesh/D3D12DrawMesh/AssetManager.h
@@ -15,7 +15,7 @@ private:
 public:
 	static FAssetManager* Get();
 	shared_ptr<FScene> LoadScene(const std::wstring& BinFileName);
-	FStaticMeshActor CreateMeshActor(uint32 VertexStride, vector<float> Vertices, vector<uint32> Indices, FTransform Transform);
+	FStaticMeshActor CreateMeshActor(uint16 VertexStride, vector<float> Vertices, vector<uint32> Indices, FTransform Transform);
 
 private:
 	void ReadMeshLODFromIfstream(std::ifstream& Fin, FStaticMeshLOD& MeshLOD);
