+++ D3D12DrawMesh/D3D12DrawMesh/StaticMeshActor.h
@@ -17,7 +17,7 @@ public:
 	const vector<float>& GetVertices() const { return Vertices; }
 	const vector<uint32>& GetIndices() const { return Indices; }
 	const uint32& GetVertexStride() const { return VertexStride; }
-	void SetVertexStride(const uint32& Stride) { VertexStride = Stride; }
+	void SetVertexStride(const uint16& Stride) { VertexStride = Stride; }
 	void SetVertices(const vector<float>& Param) { Vertices = Param; }
 	void SetIndices(const vector<uint32>& Param) { Indices = Param; }
 };
